+++ doc/CHANGES.rst
@@ -21,6 +21,8 @@ CHANGELOG
   none / empty value.
 * [FIX] option ``install_or_update_bloks``, if one blok is marked as 
   **toupdate**, then this blok will be not marked as **toinstall**
+* [FIX] ``Blok.import_file`` have not raise if does not found 
+  **error_found** in the result of the import
 
 0.17.2 (2018-02-27)
 -------------------