+++ doc/CHANGES.rst
@@ -23,6 +23,7 @@ CHANGELOG
   **toupdate**, then this blok will be not marked as **toinstall**
 * [FIX] ``Blok.import_file`` have not raise if does not found 
   **error_found** in the result of the import
+* [FIX] ``Model.get_primary_keys`` with polymorphism (single table)
 
 0.17.2 (2018-02-27)
 -------------------